fix for issue #93
- moved functions webview_js_encode and webview_inject_css into webview.h
