+++ webview-sys/webview.h
@@ -13,6 +13,8 @@ extern "C" {
 
 #include <stdint.h>
 #include <string.h>
+#include <stdlib.h>
+#include <stdio.h>
 
 typedef void* webview_t;
 typedef void (*webview_external_invoke_cb_t)(webview_t w, const char *arg);
@@ -78,6 +80,44 @@ struct webview_dispatch_arg {
   "getElementsByTagName('head')[0];t.setAttribute('type','text/"               \
   "css'),t.styleSheet?t.styleSheet.cssText=e:t.appendChild(document."          \
   "createTextNode(e)),d.appendChild(t)})"
+
+static int webview_js_encode(const char *s, char *esc, size_t n) {
+  int r = 1; /* At least one byte for trailing zero */
+  for (; *s; s++) {
+    const unsigned char c = *s;
+    if (c >= 0x20 && c < 0x80 && strchr("<>\\'\"", c) == NULL) {
+      if (n > 0) {
+        *esc++ = c;
+        n--;
+      }
+      r++;
+    } else {
+      if (n > 0) {
+        snprintf(esc, n, "\\x%02x", (int)c);
+        esc += 4;
+        n -= 4;
+      }
+      r += 4;
+    }
+  }
+  return r;
+}
+
+WEBVIEW_API int webview_inject_css(webview_t w, const char *css) {
+  int n = webview_js_encode(css, NULL, 0);
+  char *esc = (char *)calloc(1, sizeof(CSS_INJECT_FUNCTION) + n + 4);
+  if (esc == NULL) {
+    return -1;
+  }
+  char *js = (char *)calloc(1, n);
+  webview_js_encode(css, js, n);
+  snprintf(esc, sizeof(CSS_INJECT_FUNCTION) + n + 4, "%s(\"%s\")",
+           CSS_INJECT_FUNCTION, js);
+  int r = webview_eval(w, esc);
+  free(js);
+  free(esc);
+  return r;
+}
 
 static inline const char *webview_check_url(const char *url) {
   if (url == NULL || strlen(url) == 0) {
