+++ webview-sys/webview_edge.cpp
@@ -588,28 +588,6 @@ static inline char *webview_from_utf16(WCHAR *ws) {
   return s;
 }
 
-static int webview_js_encode(const char *s, char *esc, size_t n) {
-  int r = 1; /* At least one byte for trailing zero */
-  for (; *s; s++) {
-    const unsigned char c = *s;
-    if (c >= 0x20 && c < 0x80 && strchr("<>\\'\"", c) == NULL) {
-      if (n > 0) {
-        *esc++ = c;
-        n--;
-      }
-      r++;
-    } else {
-      if (n > 0) {
-        snprintf(esc, n, "\\x%02x", (int)c);
-        esc += 4;
-        n -= 4;
-      }
-      r += 4;
-    }
-  }
-  return r;
-}
-
 #include <shobjidl.h>
 
 #ifdef __cplusplus
@@ -721,22 +699,6 @@ WEBVIEW_API int webview_eval(webview_t w, const char *js)
 {
     static_cast<webview::webview*>(w)->eval(js);
     return 0;
-}
-
-WEBVIEW_API int webview_inject_css(webview_t w, const char *css) {
-  int n = webview_js_encode(css, NULL, 0);
-  char *esc = (char *)calloc(1, sizeof(CSS_INJECT_FUNCTION) + n + 4);
-  if (esc == NULL) {
-    return -1;
-  }
-  char *js = (char *)calloc(1, n);
-  webview_js_encode(css, js, n);
-  snprintf(esc, sizeof(CSS_INJECT_FUNCTION) + n + 4, "%s(\"%s\")",
-           CSS_INJECT_FUNCTION, js);
-  int r = webview_eval(w, esc);
-  free(js);
-  free(esc);
-  return r;
 }
 
 WEBVIEW_API void webview_set_title(webview_t w, const char *title)
